+++ core/commons/networks/public_ip_tracker.go
@@ -33,13 +33,25 @@ func NewFreeIpAddrTracker(subnetMask string, alreadyTakenIps []string) (ipAddrTr
 	return ipAddrTracker, nil
 }
 
+// TODO rework this entire function to handle IPv6 as well (currently breaks on IPv6)
 func (networkManager FreeIpAddrTracker) GetFreeIpAddr() (ipAddr string, err error){
 	// convert IPNet struct mask and address to uint32
 	// network is BigEndian
 	mask := binary.BigEndian.Uint32(networkManager.subnet.Mask)
 
+	subnetIp := networkManager.subnet.IP
+
+	// The IP can be either 4 bytes or 16 bytes long; we need to handle both!
+	// See https://gist.github.com/ammario/649d4c0da650162efd404af23e25b86b
+	var intIp uint32
+	if len(subnetIp) == 16 {
+		intIp = binary.BigEndian.Uint32(subnetIp[12:16])
+	} else {
+		intIp = binary.BigEndian.Uint32(subnetIp)
+	}
+
 	// We remove the zeroth IP because it's only used for specifying the network itself
-	start := binary.BigEndian.Uint32(networkManager.subnet.IP) + 1
+	start := intIp + 1
 
 	// find the final address
 	finish := (start & mask) | (mask ^ 0xffffffff)